% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_hybrid.R
\name{compute_next_generation}
\alias{compute_next_generation}
\title{computes the next fuzzycoco generation from the current one}
\usage{
compute_next_generation(engine)
}
\value{
the objective value for this new generation
}
\description{
computes the next fuzzycoco generation from the current one
}
