% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{compute_optimal_quantile_fuzzy_set_positions}
\alias{compute_optimal_quantile_fuzzy_set_positions}
\title{computes the optimal fuzzy set positions based on the distribution of the data}
\usage{
compute_optimal_quantile_fuzzy_set_positions(df, nb_sets)
}
\arguments{
\item{df}{the data as a data frame}

\item{nb_sets}{the number of fuzzy sets}
}
\value{
a list, named after the df column names, holding the vector of positions per variable
}
\description{
computes the optimal fuzzy set positions based on the distribution of the data
}
\examples{
pos <- compute_optimal_quantile_fuzzy_set_positions(mtcars, 3)
print("position for 2nd fuzzy set for qsec var", pos$qsec[[2]])
}
