% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_hybrid.R
\name{describe_best_system}
\alias{describe_best_system}
\title{describes the best fuzzy system evaluated so far}
\usage{
describe_best_system(engine)
}
\value{
the best fuzzy system as a named list
}
\description{
describes the best fuzzy system evaluated so far
}
