% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_hybrid.R
\name{describe_current_generation}
\alias{describe_current_generation}
\title{describes the current generation}
\usage{
describe_current_generation(engine)
}
\value{
the current generation as a named list
}
\description{
describes the current generation
}
