% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_iris36}
\alias{example_iris36}
\title{model parameters and data for the IRIS36 classification example}
\usage{
example_iris36()
}
\value{
the example as a named list with:
\itemize{
\item params: the model parameters
\item data: the data to fit as a data frame
}
}
\description{
a small (36 rows) dataset extracted from iris with a binary 0/1 outcome \code{OUT} response variable
}
\examples{
 model <- fuzzycoco("classification", example_iris36()$params, seed = 123)
 fit <- fit(model, OUT ~ ., example_iris36()$data, 
 max_generations = 20, progress = FALSE)
}
