% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_mtcars}
\alias{example_mtcars}
\title{model parameters and data for the mtcars regression example}
\usage{
example_mtcars()
}
\value{
the example as a named list with:
\itemize{
\item params: the model parameters
\item data: the data to fit as a data frame
}
}
\description{
model parameters and data for the mtcars regression example
}
\examples{
 model <- fuzzycoco("regression", example_mtcars()$params)
 fit <- fit(model, qsec ~ ., example_mtcars()$data, max_generations = 20, progress = FALSE)
}
