% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_hybrid.R
\name{get_current_generation_fitness}
\alias{get_current_generation_fitness}
\title{get the current generation fitness}
\usage{
get_current_generation_fitness(engine)
}
\value{
the fitness as a numeric
}
\description{
get the current generation fitness
}
