% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_hybrid.R
\name{get_current_generation_nb}
\alias{get_current_generation_nb}
\title{get the current generation number}
\usage{
get_current_generation_nb(engine)
}
\value{
the number as an integer
}
\description{
get the current generation number
}
