% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_fuzzy_coco_eval}
\alias{rcpp_fuzzy_coco_eval}
\title{FuzzyCoco::eval()}
\usage{
rcpp_fuzzy_coco_eval(df, fuzzy_system_desc, params, verbose = FALSE)
}
\value{
a data.frame
}
\description{
FuzzyCoco::eval()
}
