% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_fuzzy_coco_predict}
\alias{rcpp_fuzzy_coco_predict}
\title{FuzzyCoco::loadAndPredict()}
\usage{
rcpp_fuzzy_coco_predict(df, fuzzy_system_desc, verbose = FALSE)
}
\value{
a data.frame
}
\description{
FuzzyCoco::loadAndPredict()
}
