% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.fuzzycoco_fit}
\alias{evaluate.fuzzycoco_fit}
\title{evaluate the fuzzy system from a fit on some given data}
\usage{
\method{evaluate}{fuzzycoco_fit}(x, data, verbose = FALSE, ...)
}
\arguments{
\item{x}{the fuzzycoco_fit object containing the fuzzy system to evaluate}

\item{data}{the data to evaluate the fuzzy system on}

\item{verbose}{whether to be verbose}

\item{...}{not used. Only for S3 generic consistency}
}
\value{
the evaluation as a named list
}
\description{
N.B: just a S3 method method wrapping the \code{\link[=evaluate_fuzzy_system]{evaluate_fuzzy_system()}} function
}
\examples{
pms <- params(
 nb_rules = 2, nb_max_var_per_rule = 3,        # structural parameters
 rules.pop_size = 100, mfs.pop_size = 100,     # coevolution population sizes
 ivars.nb_sets = 3, , ivars.nb_bits_pos = 8,   # input vars: 3 fuzzy sets, and 8 bits to discretize the values 
 ovars.nb_sets = 3, ovars.nb_bits_pos = 8,     # output vars: 3 fuzzy sets, and 8 bits to discretize the values 
 metricsw.sensitivity = 0, metricsw.specificity = 0, metricsw.rmse = 1, # we just use RMSE (root mean square error)
 output_vars_defuzz_thresholds = 17            # threshold for the qsec output variable
)
model <- fuzzycoco("regression", pms, seed = 123)
df <- mtcars[c("mpg", "hp", "wt", "qsec")]
fit <- fit(model, qsec ~ ., df, engine = "rcpp", seed = 456, max_generations = 20)

res <- evaluate(fit, df)
print(res$fitness)
}
