% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.fuzzycoco_fit}
\alias{evaluate.fuzzycoco_fit}
\title{evaluate the fuzzy system from a fit on some given data}
\usage{
\method{evaluate}{fuzzycoco_fit}(x, data, verbose = FALSE, ...)
}
\arguments{
\item{x}{the fuzzycoco_fit object containing the fuzzy system to evaluate}

\item{data}{the data to evaluate the fuzzy system on}

\item{verbose}{whether to be verbose}

\item{...}{not used. Only for S3 generic consistency}
}
\value{
the evaluation as a named list
}
\description{
N.B: just a S3 method method wrapping the \code{\link[=evaluate_fuzzy_system]{evaluate_fuzzy_system()}} function
}
