% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_fuzzy_system}
\alias{evaluate_fuzzy_system}
\title{evaluate the fuzzy system from a fit on some given data}
\usage{
evaluate_fuzzy_system(fs, data, params, verbose = FALSE)
}
\arguments{
\item{fs}{the fuzzy system to evaluate (as a named list)}

\item{data}{the data to evaluate the fuzzy system on}

\item{params}{the fuzzycoco parameters. probably not needed...}

\item{verbose}{whether to be verbose}
}
\value{
the evaluation as a named list
}
\description{
evaluate the fuzzy system from a fit on some given data
}
