% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_fuzzy_system}
\alias{evaluate_fuzzy_system}
\title{evaluate the fuzzy system from a fit on some given data}
\usage{
evaluate_fuzzy_system(fs, data, params, verbose = FALSE)
}
\arguments{
\item{fs}{the fuzzy system to evaluate (as a named list)}

\item{data}{the data to evaluate the fuzzy system on}

\item{params}{the fuzzycoco parameters. probably not needed...}

\item{verbose}{whether to be verbose}
}
\value{
the evaluation as a named list
}
\description{
evaluate the fuzzy system from a fit on some given data
}
\examples{
pms <- params(
 nb_rules = 2, nb_max_var_per_rule = 3,        # structural parameters
 rules.pop_size = 100, mfs.pop_size = 100,     # coevolution population sizes
 ivars.nb_sets = 3, , ivars.nb_bits_pos = 8,   # input vars: 3 fuzzy sets, and 8 bits to discretize the values 
 ovars.nb_sets = 3, ovars.nb_bits_pos = 8,     # output vars: 3 fuzzy sets, and 8 bits to discretize the values 
 metricsw.sensitivity = 0, metricsw.specificity = 0, metricsw.rmse = 1, # we just use RMSE (root mean square error)
 output_vars_defuzz_thresholds = 17            # threshold for the qsec output variable
)
model <- fuzzycoco("regression", pms, seed = 123)
x <- mtcars[c("mpg", "hp", "wt")]
y <- mtcars["qsec"]
fit <- fit_xy(model, x, y, progress = FALSE)

res <- evaluate_fuzzy_system(fit$fuzzy_system, cbind(x, y), fit$params)
print(res$metrics$rmse)

}
