% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example_iris_binary_categorical}
\alias{example_iris_binary_categorical}
\title{model parameters and data for the IRIS36 classification example}
\usage{
example_iris_binary_categorical()
}
\value{
the example as a named list with:
\itemize{
\item params: the model parameters
\item data: the data to fit as a data frame
}
}
\description{
an example dataset based on iris with a binary categorical (non-factor) response
}
\examples{
 model <- fuzzycoco("classification", example_iris_binary_categorical()$params)
 fit <- fit(model, qsec ~ ., example_iris_binary_categorical()$data, max_generations = 20, progress = FALSE)
}
