% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_utils.R
\name{fs_rules_to_df}
\alias{fs_rules_to_df}
\title{format the fuzzy rules as a data frame}
\usage{
fs_rules_to_df(fuzzy_system_desc)
}
\arguments{
\item{fuzzy_system_desc}{a fuzzy system description as a named list}
}
\value{
a data frame, one row per rule, including the default rule, in columns the input and ouput variables.
The values are the corresponding fuzzy set number.
}
\description{
format the fuzzy rules as a data frame
}
\seealso{
Other fit_utils: 
\code{\link{fit_to_df}()},
\code{\link{fs_used_vars_to_df}()}
}
\concept{fit_utils}
