% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\name{fuzzy_coco_parsnip}
\alias{fuzzy_coco_parsnip}
\title{parsnip model function}
\usage{
fuzzy_coco_parsnip(
  mode = "unknown",
  params,
  engine = FUZZY_COCO_HYBRID_ENGINE,
  seed = sample.int(10^5, 1),
  verbose = FALSE
)
}
\description{
parsnip model function
}
