% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\name{fuzzy_coco_parsnip_wrapper}
\alias{fuzzy_coco_parsnip_wrapper}
\title{this is an utility function used to implement the parsnip interface}
\usage{
fuzzy_coco_parsnip_wrapper(
  formula,
  data,
  object = NULL,
  internal_model = NULL,
  engine = NULL,
  fit = NULL,
  df = NULL,
  pred = FALSE,
  type = NULL,
  ...
)
}
\arguments{
\item{formula}{the fuzzy coco model as a formula}

\item{data}{the data to fit as a data frame.  The output variables must be grouped AFTER the input variables}

\item{object, internal_model, fit, df, pred, type, ...}{no comment}

\item{engine}{the fuzzy coco fit engine to use, one of \strong{rcpp} and \strong{hybrid}}
}
\description{
It should not be exported, it only is because of parsnip internal implementation
}
\keyword{internal}
