% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_hybrid.R
\name{stop_engine_if_stalling}
\alias{stop_engine_if_stalling}
\title{an utility function to easily generate a stop function that stops when the convergence is stalling}
\usage{
stop_engine_if_stalling(nb_iterations, verbose = FALSE)
}
\arguments{
\item{nb_iterations}{number of iterations of the stalling: stops if the fitness has not increased during that
number of iterations.}

\item{verbose}{whether to be verbose}
}
\value{
a function engine --> logical that stops (i.e/ returns TRUE) if the convergence is stalling
}
\description{
an utility function to easily generate a stop function that stops when the convergence is stalling
}
\examples{
until <- stop_engine_on_first_of(max_generations = 1000, max_fitness = 0.9, other_func = stop_engine_if_stalling(5))
}
